[6주차/빈] 워크북 제출합니다

+++ src/main/java/umc/domain/member/dto/MemberResDTO.java
@@ -7,7 +7,7 @@ public class MemberResDTO {
     //마이페이지
     @Builder
     public record MyPageResDTO(
-            int memberId,
+            Long memberId,
             String name,
             String email,
             String phoneNumber,
